move advance_chain to a new trait AdvanceChain, impl for AuthProvider, cleanup types in spammer mod
