+++ crates/core/src/eth_engine/mod.rs
@@ -1,7 +1,9 @@
-mod auth_provider;
+mod auth_provider_eth;
 mod auth_transport;
+mod traits;
 mod util;
 mod valid_payload;
 
-pub use auth_provider::AuthProvider;
+pub use auth_provider_eth::AuthProvider;
+pub use traits::AdvanceChain;
 pub use util::*;
