+++ crates/core/src/eth_engine/traits.rs
@@ -1,0 +1,7 @@
+use async_trait::async_trait;
+
+#[async_trait]
+pub trait AdvanceChain {
+    /// Advance the chain by calling `engine_forkchoiceUpdated` (FCU) and `engine_newPayload` methods.
+    async fn advance_chain(&self, block_time_secs: u64) -> Result<(), Box<dyn std::error::Error>>;
+}
